pyproject.toml
[tool.poetry]
name = "drf-test"
version = "0.1.0"
description = ""
authors = ["Alex Kivaiko <alexagree1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
django-polymorphic = "^3.1.0"
django-debug-toolbar = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80

[tool.isort]
include_trailing_comma = true
line-length = 80
line_after_imports = 2
skip = "__init__.py"

[tool.flake8]
exclude = ".git,__pycache__,__init__.py"
max-line-length = 80
ignore = ["E203", "E266", "E501", "W503"]
